refactor: extract UriLocation::from_uri.
